harvest-dwarf: read DW_AT_frame_base rather than assuming it
DW_TAG_subprogram contains frame-base which is implicitly referred in
DW_OP_fbreg.  Previously we assuming it to be equal to iesp which is
true for LLVM but not for GCC.  Now we read it from DW_AT_frame_base.

Slightly hacky: DW_OP_call_frame_cfa is assumed to be
(iesp + ADDRESS_SIZE_IN_BYTES)

Re-organized the code somewhat as well.