llvm2tfg: handle address taken params and expanded struct params
As LLVM IR accepts params as values, for an address taken parameter
Clang alloca'tes a new region, stores the parameter value in it,
and uses it in rest of the program (instead of the passed-by-value
param).

Clang does similar thing for a struct param where its members are
expanded in the parameter list, a struct and initialized alloca'ted in
the prologue of the function and used instead of the passed parameter.

While the binary generated by Clang/LLVM does the same thing for expanded
struct param (i.e. do allocation and use it), for the address taken
parameter it seemingly reuses the address of the passed parameter
(instead of doing a fresh allocation).

Both GCC and ICC, on the other hand, seemingly always use the passed
parameters and do not do fresh allocation.

In summary, with the new change:

* Benchmark with address taken scalar parameter passes for all 3
compilers
* Benchmark with struct as passed-by-value parameter (which is expanded
due to ABI requirements) does/will not pass for clang.

A simple fix for the clang issue could be passing an additional flag
for choosing the behavior for handling expanded structs.